requirements.txt
discord.py==2.3.2
groq==0.11.0
motor==3.5.1
pymongo==4.8.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
